Split off ChecklistsController, sort out bonuses inflections
