+++ app/views/checklist/show.html.erb
