+++ app/views/observer/show_site_stats.html.erb
@@ -3,7 +3,7 @@
 
   tabs = [
     link_to(:app_contributors.t, users_by_contribution_path),
-    link_to(:site_stats_observed_taxa.t, users_checklist_path)
+    link_to(:site_stats_observed_taxa.t, checklist_path)
   ]
   @tabsets = { right: draw_tab_set(tabs) }
 %>
