+++ app/views/species_list/show_species_list.html.erb
@@ -84,7 +84,7 @@
             <%= link_to_coerced_query(@query, Name) %> |
             <%= link_to_coerced_query(@query, Image) %> |
             <%= link_to(:app_checklist.t,
-                        { users_checklist_path(
+                        { checklist_path(
                           species_list_id: @species_list.id
                           ) }) %> |
             <%= link_to(:show_object.t(type: :map),
